refactor: Extract boxing.rs
